% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc.tube.R
\name{calc.tube}
\alias{calc.tube}
\alias{calcTubeStat}
\title{Common Diffusion Tube Calculations}
\usage{
calcTubeStat(data, tube = ".value", by = NULL, stat = NULL, ...)
}
\arguments{
\item{data}{Data source, typically a data.frame or similar, containing
data-series of diffusion tube records.}

\item{tube}{The name of the data-series (in \code{data}) to calculate,
the requested statistic (or statistics) for (see \code{stat}),
typically the DT NO2 concentrations (in ug/m3).}

\item{by}{The name(s) of any grouping terms. If none are supplied,
total sample statistic(s) are calculated for \code{data}.}

\item{stat}{The statistic to calculate, by default the mean in the form:

\code{mean(x, na.rm=TRUE)}

But it can be a list of one or more functions, ideally supplied in
the form:

\code{function(x){list(mean=mean(x, na.rm), sd=sd(x, na.rm=TRUE))}}}

\item{...}{additional arguments, currently ignored.}
}
\value{
All functions return a \code{data.frame} of requested
\code{data} statistics.
}
\description{
Functions for common diffusion tube (DT)
data calculations using \code{DTEval}.
}
\details{
\code{calcTubeStat} attempts to calculate summary statistics.

It and related functions assume that \code{data} is data
\code{DTEval} will recognise as Diffusion Tube data, so either
previously tagged tube data or data that is tag-able using a
default call of \code{\link{tagTube}}
}
