% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caz.brd.R
\docType{data}
\name{caz.brd}
\alias{caz.brd}
\title{caz.brd}
\format{
A (1x2) 'sf' object
\describe{
\item{name}{File description, 'Bradford CAZ 2021 MULTILINESTRING'.}
\item{geom}{Associated GEOM data}
}
}
\source{
Bradford CAZ borders as 'sf' object, as supplied.
}
\usage{
caz.brd
}
\description{
Clean Air Zone (CAZ) borders for the 2022 Bradford CAZ.
}
\keyword{datasets}
