% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deseason.tube.R
\name{deseason.tube}
\alias{deseason.tube}
\alias{deseasonTubeData}
\title{Deseasonalise Diffusion Tube Data}
\usage{
deseasonTubeData(data, tube = ".value", by = NULL, ...)
}
\arguments{
\item{data}{Data source, typically a data.frame or similar, containing
data-series of diffusion tube records.}

\item{tube}{The name of the data-series (in \code{data}) to fit a model
to (see \code{model}), typically the DT NO2 concentrations (in ug/m3).}

\item{by}{The name(s) of hierarchical grouping terms. These are used to
sub-sample the data when deseasonalising data.}

\item{...}{additional arguments, currently ignored.}
}
\value{
All functions return the supplied \code{data.frame}
with attached predicted time-series components: \code{..season},
\code{..deseason}, etc.
}
\description{
Functions for seasonalising multiple year
diffusion tube (DT) data time-series.
}
\details{
\code{deseasonaliseTube} attempts to deseasonalise the supplied time-series.

It builds a LOESS time-series model in the form:

\code{[tube] ~ loess([day-of-year] + [date])}

And, extracts the seasonal component as response term for the day-of-year.
}
\note{
\code{deseasonTubeData} and related functions assume that \code{data}
is a data set \code{DTEval} will recognise as Diffusion Tube data, so either
previously tagged tube data or data that is tag-able using a
default call of \code{\link{tagTube}}
}
