% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dt.brd.R
\docType{data}
\name{dt.brd}
\alias{dt.brd}
\title{dt.brd}
\format{
A (11273x11) 'data.frame' object
\describe{
\item{site}{Site identifier code}
\item{site_name}{Site name/description}
\item{year_of_measurement}{Year the measurement was made (numeric, YYYY)}
\item{month}{Month the measurement was made (character, 3-letter abbreviation)}
\item{month_numeric}{Month the measurement was made (numeric, 1-12)}
\item{latitude}{Latitude of site}
\item{longitude}{Longitude of site}
\item{local_authority}{Local Authority making these measurements}
\item{bias_adjusted_measurement}{The bias adjusted NO2 measurement}
\item{.start_date}{Sampling start date (Date, YYYY-MM-DD)}
\item{.end_date}{Sampling end date (Date, YYYY-MM-DD)}
}
}
\source{
Bradford City Council
}
\usage{
dt.brd
}
\description{
Bradford diffusion tube data set, 2020 to 2025 subset of
tube measurements collected at sites classified as valid for
roadside sampling using the AQSR sampling scheme.
}
\note{
Start and end dates of individual diffusion tube sampling periods
(\code{.start_date} and \code{.end_date}) based of Defra Sampling
Calendar, see \link{\code{dt.calendar}}.
}
\examples{
head(dt.brd)

}
\keyword{datasets}
