% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dt.calendar.R
\docType{data}
\name{dt.calendar}
\alias{dt.calendar}
\title{dt.calendar}
\format{
A (132x5) 'data.frame' object
\describe{
\item{year}{sampling year}
\item{month}{sampling month}
\item{start}{start date for sampling that year and month}
\item{end}{end date for sampling that year and month}
\item{weeks}{number of weeks sampled}
}
}
\source{
These were extracted from yearly sampling schedules at (or from):

https://laqm.defra.gov.uk/diffusion-tubes/data-entry (2015-2020),

https://laqm.defra.gov.uk/air-quality/air-quality-assessment/diffusion-tube-monitoring-calendar/ (2021),

LAQM Helpdesk Team (2022),

https://laqm.defra.gov.uk/air-quality/air-quality-assessment/diffusion-tube-monitoring-calendar/ (2023-2025).
}
\usage{
dt.calendar
}
\description{
The diffusion tube sampling schedule calendar
2015 and 2025.
}
\keyword{datasets}
