% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DTEval-package.R
\name{DTEval-package}
\alias{DTEval-package}
\alias{DTEval}
\alias{"_PACKAGE"}
\title{DTEval: Diffusion Tube Evaluation Code}
\description{
R code used for the pro-processing, analysis and evaluation of diffusion tube (DT)
data typically collected in air quality assessment exercises.
}
\details{
This package was splat packed.
}
\section{Main Package Functions}{

\code{\link{testTubeAccuracy}}, \code{\link{testTubePrecision}}
}

