% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.tube.R
\name{fit.tube}
\alias{fit.tube}
\alias{fitTubeModel}
\alias{fitTubeModel_gam}
\alias{fitTubeModel_loess}
\title{Fit Diffusion Tube Models}
\usage{
fitTubeModel(
  data,
  tube = ".value",
  inputs = NULL,
  by = NULL,
  model = NULL,
  simplify = FALSE,
  min.count = -1,
  min.prop = -1,
  new.data = NULL,
  ...
)

fitTubeModel_gam(data, tube, inputs, new.data = NULL, ...)

fitTubeModel_loess(data, tube, inputs, new.data = NULL, ...)
}
\arguments{
\item{data}{Data source, typically a data.frame or similar, containing
data-series of diffusion tube records.}

\item{tube}{The name of the data-series (in \code{data}) to fit a model
to (see \code{model}), typically the DT NO2 concentrations (in ug/m3).}

\item{inputs}{The name(s) of the model inputs.}

\item{by}{The name(s) of hierarchical grouping terms. These are used to
sub-sample the data when build multiple models.}

\item{model}{The model to fit, by default \code{fitTubeModel_loess}}

\item{simplify}{Option to simplify the \code{data} before fitting by
averaging common inputs, default \code{FALSE}.}

\item{min.count}{The minimum data count required for a model build.
The default, -1, disables this option.}

\item{min.prop}{The minimum proportion of non-NA data required for a model
build. The default, -1, disables this option and values higher than 1
exclude all supplied data.}

\item{new.data}{The type of prediction to generate. The default,
\code{NULL}, returns the supplied \code{data} with an extra column
of the requested predictions added as \code{[tube].pred}. The alternative
option \code{'input.ranges'} generates a \code{data.frame} of the model
input ranges.}

\item{...}{additional arguments, currently passed on to \code{model} and
used if recognised.}
}
\value{
All functions return a \code{data.frame} of requested
\code{data} statistics.
}
\description{
Functions for fitting diffusion tube (DT)
data calculations using \code{DTEval}.
}
\details{
\code{fitTubeModel} attempts to fit a model to the supplied inputs.

It and related functions assume that \code{data} is data
\code{DTEval} will recognise as Diffusion Tube data, so either
previously tagged tube data or data that is tag-able using a
default call of \code{\link{tagTube}}
}
\examples{
\dontrun{
#basic examples
mod <- fitTubeModel(dt.brd, inputs=c(".latitude", ".longitude"),
                    by =".year", simplify = T)
tubePlot(mod, ".value", ".value.pred", facet=".year")
}


}
