% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.dt.date.R
\name{misc.dt.date}
\alias{misc.dt.date}
\alias{setTubeDate}
\title{Miscellaneous Date Handlers}
\usage{
setTubeDate(data, method = 2, ...)
}
\arguments{
\item{data}{Data source, typically a data.frame or similar, containing
data-series of diffusion tube records.}

\item{method}{The method to use (see details).}

\item{...}{additional arguments, currently ignored.}
}
\value{
By default \code{setTubeDate} returns \code{data} with an
additional column, \code{.date} of sampling dates determined using
requested method.
}
\description{
Miscellaneous functions for use with diffusion tube date
information in \code{DTEval}.

Diffusion tubes are typically deployed for
several weeks, so each sample record has an associated sampling start and
end date (\code{.start_date} and \code{.end_date} once tagged, see
\code{\link{tagTube}}) and an associated sampling interval. These functions
are intended as shortcuts for those unfamiliar with the \code{R} \code{Date}
object class used by \code{DTEval}.
}
\details{
\code{setTubeDate} attempts to assign a representative date to each
sampling record (\code{data} row). By default, it set the date to the
mid-point between sample start and end dates, and is, for example,
intended for use when plotting time-series: options: 1 (start date),
2 (default; mid-point date); 3 (end date).

All functions assume \code{data} is either  previously tagged data or data
that can be tagged using a default call of \code{\link{tagTube}}.
}
