% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.dt.handlers.R
\name{misc.dt.handlers}
\alias{misc.dt.handlers}
\alias{getTubeX}
\alias{checkTubeData}
\title{Miscellaneous Diffusion Tube Data Handling}
\usage{
getTubeX(data, x = NULL, ..., test.class = NULL, if.err = "return.null")

checkTubeData(data, x = NULL, ..., n.x = -1, if.err = "return.null")
}
\arguments{
\item{data}{Data source, typically a data.frame or similar, containing
data-series of diffusion tube records.}

\item{x}{The name of a data-series in \code{data} or expression to
be evaluated, supplied as a character string. See Details below.}

\item{...}{additional arguments, currently ignored.}

\item{test.class}{If supplied, the required class of \code{x} as a
character string.}

\item{if.err}{The action to take if the function fails: By default,
\code{'return.null'} to return a \code{NULL} silently, but other
options include \code{'stop'} to stop as \code{getTubeX},
and \code{'stop<<FUN>>ARG'} to stop with function 'FUN' reporting a
problem with argument ARG.}

\item{n.x}{The maximum number of \code{x} that to be checked.}
}
\value{
\code{getTubeX} returns \code{x} if it can
be evaluated with the supplied \code{data}.

\code{getTubeData} returns \code{data} plus any additions columns required to
allow all valid \code{x} to be used directly as column names, e.g, in form:

\code{data[, x]}
}
\description{
Miscellaneous code for routine diffusion tube (DT) data
handling with \code{DTEval}.
}
\details{
\code{getTubeX} attempts to extract \code{x} from \code{data} (or
build it from information in \code{data}. It uses \code{with(data, x)},
so can evaluates terms like:

\code{getTubeX(data, 'factor(y)')}

(assuming \code{y} is in \code{data} or visible from it, and \code{x}
is a character string)

It is intended for use with a single \code{x} term and returns it as a
vector.

\code{checkTubeData} is a general purpose wrapper for \code{getTubeX} that
handles multiple \code{x} terms. It attempts to evaluate each in turn and
returns the supplied \code{data} plus any additions required.
}
