% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.dt.lat.lon.R
\name{misc.dt.lat.lon}
\alias{misc.dt.lat.lon}
\alias{tubeInXYPolygon}
\title{Miscellaneous Latitude and Longitude Handlers##}
\usage{
tubeInXYPolygon(data, polygon, ...)
}
\arguments{
\item{data}{Data source, typically a data.frame or similar, containing
data-series of diffusion tube records.}

\item{polygon}{Data source or polygon shape file.}

\item{...}{additional arguments, currently ignored.}
}
\value{
By default \code{tubeInXYPolygon} returns \code{data} with an
additional column, \code{.in_polygon} that tube inclusion as \code{TRUE}
or \code{FALSE}.
}
\description{
Miscellaneous functions for use with Latitude and Longtitude
data. By default, these are setup for use with \code{DTEval}, so
expect tagged data (see \code{\link{tagTube}}).
}
\details{
\code{TubeinXYPolygon} attempts to identify all records in \code{data},
that are within the boundaries of the supplied \code{polygon}. It is a
wrapper for \code{splancs} function \code{inout} that uses the latitude
and longitude tags of the supplied \code{data}, and the \code{x} and
\code{y} columns of the supplied \code{polygon} to test \code{data} rows
for inclusion. However, if you want to use different \code{data} or
\code{polygon} columns these can be reset using \code{lat} and \code{lon}
for the \code{data}-point locations and \code{x} and \code{y} for
\code{polygon}.

All functions assumed \code{data} and \code{polygon} contain point and
polygon source information, that \code{data} is data \code{DTEval} will
recognise as diffusion tube data, so either previously tagged tube data
or data that is tag-able using a default call of \code{\link{tagTube}},
and that these and the x/y data in \code{polygon} are WGS84 format
map coordinates.
}
