% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.dt.meta.R
\name{misc.dt.meta}
\alias{misc.dt.meta}
\alias{addTubeMeta}
\alias{extractTubeMeta}
\alias{padTubeMeta}
\title{Miscellaneous Diffusion Tube Meta Data}
\usage{
addTubeMeta(data, ref = NULL, by = NULL, ...)

extractTubeMeta(data, x = NULL, by = NULL, ...)

padTubeMeta(data, x = NULL, by = NULL, ...)
}
\arguments{
\item{data}{Data source, typically a data.frame or similar, containing
data-series of diffusion tube records.}

\item{ref}{(For \code{addTubeMeta} only) The meta data source, typically
a data.frame or similar, containing meta-information associated with the
diffusion tube records in \code{data}. See Details below.}

\item{by}{The name of a data-series that can be used as a
case identifier for meta information. For most diffusion tube meta data,
this is often a site identifier, e.g. a site code or name. If \code{by}
is not supplied, the function will check for known \code{DTEval}
\code{tags} and file details, but may still require user input. See
Details below.}

\item{...}{additional arguments, currently ignored.}

\item{x}{(For \code{padTubeMeta} and \code{extractTubeMeta}) The name of
a data-series in \code{data} or expression to be evaluated, supplied as
a character string. See Details below.}
}
\value{
The functions return the requested data, and are generally used in the
form:

\code{requested.data <- padTubeMeta(dt.data, "[meta.name]", "[by.name]")}

\code{addTubeMeta} returns the supplied \code{data} with \code{ref}
merged at the requested level.

\code{extractTube} returns a \code{data.frame} of data that looks like
meta-data when grouped at the requested level.

\code{padTubeMeta} returns \code{data} with the requested fix,
if it can be applied.
}
\description{
Miscellaneous code used to work with
diffusion tube (DT) meta data.
}
\details{
\code{addTubeMeta} attempts to merge \code{data} and \code{ref} using
\code{by} as the common term. It is intended for use with reference
meta data or data extracted with \code{extractTubeMeta}.

\code{extractTube} attempts to extract data that looks like meta
information from the supplied \code{data}. Meta-data
is only be expected to unique at a specific level of aggregation, e.g.
\code{latitude} will be unique for a sample site. Although the function
default is to look for known sample identifiers, this grouping term
is set using the addition argument \code{by}. The function then extracts
all data-series with only one unique non-\code{NA} value. By default,
the function test all non-grouping data-series, but similarly
testing/extraction can be limited to specific cases using \code{x}.
Please remeber that this function extracts data-series that at the
\code{by} grouping-level looks like meta information.

\code{padTubeMeta} attempts to pad the \code{x} data-series by replacing
any \code{NA}s with the first non-\code{NA} entry from the same data-series.
It is intended for use with meta-data data-series, e.g. a \code{latitude}
(or \code{longitude column}) where value was only entered once.
}
