% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.dt.stats.R
\name{misc.dt.stats}
\alias{misc.dt.stats}
\alias{tubeSummary}
\alias{tubeSummarySample}
\title{Miscellaneous Diffusion Tube Data Statistics}
\usage{
tubeSummary(data, ...)

tubeSummarySample(data, ...)
}
\arguments{
\item{data}{Data source, typically a data.frame or similar, containing
data-series of diffusion tube records.}

\item{...}{additional arguments, currently ignored.}
}
\value{
All functions return a \code{data.frame} of requested
\code{data} statistics.
}
\description{
Miscellaneous code to calculate some common diffusion tube
(DT) outputs used by \code{DTEval}.
}
\details{
\code{tubeSummary} attempts to generate an overview summary for the
supplied \code{DTEval} diffusion tube data set \code{data}.

\code{tubeSummarySample} attempts to generate a summary of
sample-related counts and check sums for \code{data}.

They are also intended to be used in the form:

\code{tubeSummary(data)}

And all assumes that \code{data} is data \code{DTEval} will
recognise as Diffusion Tube data, so either previously tagged
tube data or data that is tag-able using a default call of
\code{\link{tagTube}}
}
