% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.tube.annual.R
\name{misc.tube.annual}
\alias{misc.tube.annual}
\alias{tubeAnnualCover}
\alias{tubeAnnualTest}
\title{Miscellaneous Annual Metrics}
\usage{
tubeAnnualCover(data, tube = ".value", by = ".location", ...)

tubeAnnualTest(data, test = NULL, by = ".location", years = NULL, ...)
}
\arguments{
\item{data}{Data source, typically a data.frame or similar, containing
data-series of diffusion tube records.}

\item{tube}{The \code{data} column containing the diffusion tube
measurements, by default \code{.value}.}

\item{by}{The \code{data} column containing the site or local codes to
be used when grouping \code{tube} values for annual statistic
calculation, by default \code{.location}.}

\item{...}{additional arguments, e.g. \code{output} and \code{rename}
which allow users to modify function outputs.}

\item{test}{(\code{tubeAnnualTest} only) A test to be evaluated based on
data-series in \code{data}, supplied as a character string.}

\item{years}{(\code{tubeAnnualTest} only) The years the test should be
applied to if not the full range of the data.}
}
\value{
By default \code{tubeAnnualCover} returns \code{data} with two
additional columns, \code{.annual.n} the numbers of days with
at least one measurement reported for that location in that year,
and \code{.annual.n} the percentage of days with at least one
measurement reported for that location in that year.

By default \code{tubeAnnualTest} returns \code{data} with one
additional column, the logical response to the supplied test, for
example a call including the argument \code{test='.annual.n>300'}
will, assuming your \code{data} contains a column called
\code{.annual.n}, generate a column, \code{TRUE} for all
case where all \code{.annual.n} entries for the same year and
location were larger than 300 or \code{FALSE} if not.
}
\description{
Miscellaneous functions for the calculation of annual
diffusion tube metrics. By default, these are setup for use with
\code{DTEval}, so expect tagged data (see \code{\link{tagTube}}),
and are intended to be applied by year and location to provide
information about local data coverage (or capture efficiency)
independent of co-locational sampling or site re-naming.
}
\details{
\code{tubeAnnualCover} attempts to calculate the annual data coverage,
based of the \code{data} \code{.start_date} and \code{.end_date} tags.
By default, it is binned and averaged by \code{.location} to reduce issues
associated with replicate sampling and site renaming, and calculates two
metrics: \code{n}, the number of days measurements have been
reported at that location in the year the sample was collected; and,
\code{pc}, the percent of the year that accounts for.

\code{tubeAnnualTest} can be used to generate logical conditioning
based upon \code{tubeAnnualCover} outputs. It uses a \code{test}
argument, supplied as a string, to generate a column of \code{TRUE},
\code{FALSE} responses, and can be applied to the full \code{data}
set, or limited to specific years using the extra argument
\code{years}.

All functions assume \code{data} is data \code{DTEval} will recognise
as diffusion tube data, i.e., previously tagged (see
\code{\link{tagTube}}).
}
