% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.tube.meta.R
\name{misc.tube.meta}
\alias{misc.tube.meta}
\alias{addTubeMeta}
\alias{checkTubeMeta}
\alias{extractTubeMeta}
\alias{padTubeMeta}
\title{Working with Diffusion Tube Meta Data}
\usage{
addTubeMeta(data, ref = NULL, by = NULL, ...)

checkTubeMeta(data, x = NULL, by = NULL, ...)

extractTubeMeta(data, x = NULL, by = NULL, ...)

padTubeMeta(data, x = NULL, by = NULL, ...)
}
\arguments{
\item{data}{Data source, typically a data.frame or similar, containing
data-series of diffusion tube records.}

\item{ref}{(For \code{addTubeMeta} only) The meta data source, typically
a data.frame or similar, containing meta-information associated with the
diffusion tube records in \code{data}. See Details below.}

\item{by}{The name of a data-series that can be used as a
case identifier for meta information. For most diffusion tube meta data,
this is often a site identifier, e.g. a site code or name. If \code{by}
is not supplied, the function will check for known \code{DTEval}
\code{tags} and data set details, but may still require user input. See
Details below.}

\item{...}{additional arguments, currently ignored.}

\item{x}{(For most \code{...TubeMeta} functions) The name of a data-series in
\code{data} or expression to be evaluated, supplied as a character string.
See Details below.}
}
\value{
These functions are generally intented to be used in the form:

\code{updated.data <- addTubeMeta(dt.data, ref, "[by.name]")}

\code{requested.data <- padTubeMeta(dt.data, "[meta.name]", "[by.name]")}

... etc

\code{addTubeMeta} returns the supplied \code{data} with \code{ref}
merged at the requested level.

\code{extractTube} returns a \code{data.frame} of data that looks like
meta-data when grouped at the requested level.

\code{padTubeMeta} returns \code{data} with the requested fix,
if it can be applied.
}
\description{
Miscellaneous code used to work with
diffusion tube (DT) meta data.
}
\details{
\code{addTubeMeta} attempts to merge \code{data} and \code{ref} using
\code{by} as the common merging-term. It is intended for use with reference
meta data or data extracted with from a reliable source using
\code{extractTubeMeta}.

\code{extractTubeMeta} attempts to extract data that looks like meta
information from the supplied \code{data}. Meta-data
is expected to unique at a specific level of aggregation, e.g. all
\code{latitude} records should be identical for a given sample site.
Although the function default is to look for known sample identifiers,
this grouping term can be set using the addition argument \code{by}.
The function then extracts all data-series with only one unique
non-\code{NA} value for case of \code{by}. By default, the function
tests all non-grouping data-series, but similarly testing/extraction
can be limited to specific cases using \code{x}. Please remember that
this function extracts data-series that at the \code{by} grouping-level
looks like meta information. This is no unambiguous test to identify a
data-series as meta-data. So, this function needs to be handled with care,
especially if you have any concerns about the quality of your data sets.

\code{padTubeMeta} attempts to pad the \code{x} data-series by replacing
any \code{NA}s with the first non-\code{NA} entry from the same data-series.
It is intended for use with meta-data data-series, e.g. a \code{latitude}
(or \code{longitude column}) where value was only entered once. Again, care
should be taken using this function.
}
