% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.dt.fixes.R
\name{misc.dt.fixes}
\alias{misc.dt.fixes}
\alias{padTubeMeta}
\title{Miscellaneous Diffusion Tube Data Fixes}
\usage{
padTubeMeta(data, x = NULL, site.id = NULL, ...)
}
\arguments{
\item{data}{Data source, typically a data.frame or similar, containing
data-series of diffusion tube records.}

\item{x}{The name of a data-series in \code{data} or expression to
be evaluated, supplied as a character string. See Details below.}

\item{...}{additional arguments, currently ignored.}
}
\value{
\code{getTubeX} returns \code{x} if it can
be evaluated with the supplied \code{data}.

\code{getTubeData} returns \code{data} plus any additions columns required to
allow all valid \code{x} to be used directly as column names, e.g, in form:

\code{data[, x]}
}
\description{
Miscellaneous code for fixing commonly reported issues with
diffusion tube (DT) data.
}
\details{
\code{getTubeX} attempts to extract \code{x} from \code{data} (or
build it from information in \code{data}. It uses \code{with(data, x)},
so can evaluates terms like:

\code{getTubeX(data, 'factor(y)')}

(assuming \code{y} is in \code{data} or visible from it, and \code{x}
is a character string)

It is intended for use with a single \code{x} term and returns it as a
vector.

\code{checkTubeData} is a general purpose wrapper for \code{getTubeX} that
handles multiple \code{x} terms. It attempts to evaluate each in turn and
returns the supplied \code{data} plus any additions required.
}
