% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.tube.R
\name{plot.tube}
\alias{plot.tube}
\alias{ggplotTubeShell}
\title{Diffusion Tube ggplot}
\usage{
ggplotTubeShell(data, x = NULL, y = NULL, ...)
}
\arguments{
\item{data}{Data source, typically a data.frame or similar, to be used to
build a plot using ggplot2.}

\item{x, y}{The names of the data-series to plot on the
X and Y axes, respectively.}

\item{...}{Additional arguments. See details below}
}
\value{
\code{ggplotTubeShell} returns a ggplot shell the user then
needs to add \code{geom}s to, to build a standard plot.

NOTES: Maybe I'm already missed \code{qplot}...
}
\description{
\code{DTEval} uses ggplot2 to generate most plots.
\code{ggplotTubeShell} builds common plot shells for many of these, and
handles some of of the generic plot control.
}
\details{
In addition to \code{data}, the main data source for plots,
\code{ggplotTubeShell} handles the following common plot arguments:
\itemize{
\item \code{xlab}, \code{ylab} The X and Y axes labels to use if
different from plot defaults.
\item \code{group} The name of the data-series to use to group
plotted data with. By default, \code{ggplotTubeShell} assumes
supplied groups are factors.
\item \code{facet} The name of the data-series to use to cut
the data by when generating multiple plot panels.
}
}
