% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.tube.accuracy.R
\name{test.tube.accuracy}
\alias{test.tube.accuracy}
\alias{testTubeAccuracy}
\title{Diffusion Tube Accuracy}
\usage{
testTubeAccuracy(
  data,
  data.ref = NULL,
  tube = "measurement",
  ref = "no2",
  method = 1,
  max.distance = 10,
  nearest.only = FALSE,
  show = c("plot", "summary.report"),
  ...
)
}
\arguments{
\item{data}{Data source, typically a data.frame or similar, containing
data-series of conventionally formatted DT records.}

\item{data.ref}{Reference data source, typically a data.frame or similar,
containing data-series of conventionally formatted CA
records not already aggregated, time aligned or added to \code{data}.}

\item{tube}{The name of the data-series (in \code{data}) to test, typically
DT NO2 concentrations (in ug/m3).}

\item{ref}{The name of the data-series (in \code{data.ref}) to compare with \code{tube}.}

\item{method}{The method to apply when aggregating \code{ref} (and
\code{data.ref}) to compare with \code{tube} (and \code{data}). There is
currently one method, set up to run with \code{openair}-format
\code{data.ref} data sets. This method aggregates and averages
\code{data.ref} at the sampling resolution of the tube data set, \code{data}, as
identified from tags built using \code{\link{tagTube}} methods, before merging
tube and reference data sets. See below or ?\code{\link{tagTube}}
for further details.}

\item{max.distance}{the maximum distance (m) between tube and reference
method locations when pairing diffusion tube and reference measurements for
comparison.}

\item{nearest.only}{A logical (default \code{FALSE}); if \code{TRUE} only
the nearest tube/ref pairings are retained.}

\item{show}{The outputs to show when returning results, by default the
plot and summary report.}

\item{...}{Additional arguments, currently passed to
\code{\link{tagTube}} to check tube tags, and \code{\link{ggplotTubeShell}}
to handle generic plotting arguments.}
}
\value{
\code{testTubeAccuracy} returns a list of test results, typically
\code{data}, the match (tube and reference) measurement pairs less than
\code{max.distance} apart based on tagged sampling dates and locations,
\code{all}, all date-based measurement pairs regardless of distance apart,
\code{plot}, and results \code{lookup} table and \code{report}.
}
\description{
Testing diffusion tube accuracy. Coded methods are all
currently based on the comparison of aligned diffusion tube (DT) and
reference method (continuous analyser, CA) measurements.
}
\details{
Both sampling dates and locations are identified/tagged using
\code{\link{tagTube}} methods. \code{\link{tagTubeStartEnd}} tags sampling
dates and \code{\link{tagTubeLatLon}} tags locations.

NOTE: \code{testTubeAccuracy} attempts to tag \code{data} using default
\code{\link{tagTube}} setting if it has not already been tagged, but it
is probably better to tag and check all data yourself before running more
detailed analysis.

\code{testTubeAccuracy} also uses \code{\link{ggplotTubeShell}} to handle
common plotting arguments, e.g. X and Y axes label handling. Plot
conditioning arguments like \code{group} and \code{facet} are also
tracked by \code{testTubeAccuracy} itself and used to subset data
when comparing \code{tube} and near-\code{ref} measurements.

When merging \code{data} and \code{data.ref}, \code{testTubeAccuracy} adds
a \code{'.ref'} suffix to data columns from \code{data.ref}, so it you want
to facet or group data by a column in \code{data.ref}, you need to include
the suffix when naming it.
}
