% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.tube.precision.R
\name{test.tube.precision}
\alias{test.tube.precision}
\alias{testTubePrecision}
\title{Diffusion Tube Precision}
\usage{
testTubePrecision(
  data,
  tube = ".value",
  n = 3,
  method = 2,
  show = c("plot", "summary.report"),
  ...
)
}
\arguments{
\item{data}{Data source, typically a data.frame or similar, containing
data-series of conventionally formatted diffusion tube records.}

\item{tube}{The name of the data-series (in \code{data}) to test, typically
DT NO2 concentrations (in ug/m3).}

\item{n}{The test compares replicate samples collected from the same
location at the same time. \code{n} sets the number of replicates required,
by default 3 for triplicate measurements.}

\item{method}{The method to apply when testing \code{tube} data. See relevant
details below for further information.}

\item{show}{The outputs to show when returning results, by default the
plot and summary report.}

\item{...}{Additional arguments, currently passed to
\code{\link{tagTube}} to check tube tags, and \code{\link{ggplotTubeShell}}
to handle generic plotting arguments.}
}
\value{
\code{testTubePrecision} returns a list of test results,
typically \code{data} all \code{n}-replicate measurement sets identified
using \code{\link{tagTube}}, \code{plot}, results \code{lookup} table
and; the (full) \code{report}.
}
\description{
Testing diffusion tube precision. Coded methods are all
currently based on the comparison of co-located replicate diffusion
tube (DT) samples.
}
\details{
\code{testTubePrecision} uses \code{\link{tagTube}}
to identify replicate samples, if not already tagged, then applies
one of the following methods to estimate the precision of cases
with \code{n} replicates:
\itemize{
\item \code{method = 1} uses \code{quantile(x, 0.05, 0.95)}
\item \code{method = 2} uses \code{mean(x) +/- qt(0.05/2, n(x)-1), lower.tail=FALSE)} *
\code{sd(x)/sqrt(n(x))}
\item \code{method = 3} used \code{mean(x) +/- 1.96} * \code{sd(x)/sqrt(n(x)))}
\item \code{method = 4} used \code{mean(x) +/- qnorm(0.975)} * \code{sd(x)}
}

NOTE: \code{testTubePrecision} attempts to tag \code{data} using default
\code{\link{tagTube}} setting if it has not already been tagged, but it
is probably better to tag and check all data yourself before running more
detailed analysis.

\code{testTubePrecision} also uses \code{\link{ggplotTubeShell}} handles
common plotting arguments, e.g. X and Y axes label handling. Plot
conditioning arguments like \code{group} and \code{facet} are also
tracked by \code{testTubePrecision} itself and used to subset data
when testing precision.
}
