% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tube.maps.R
\name{tube.maps}
\alias{tube.maps}
\alias{tubeMap}
\alias{leafletTubeMap}
\title{Diffusion Tube maps}
\usage{
tubeMap(data, x = NULL, y = NULL, previous = NULL, ...)

leafletTubeMap(data, x = NULL, y = NULL, previous = NULL, ...)
}
\arguments{
\item{data}{Data source, typically a data.frame or similar, to be used to
build a plot using ggplot2.}

\item{x, y}{The names of the data-series to plot on the
X and Y axes, respectively, and assumed to be sampling point longitude and
latitude.}

\item{previous}{(Optional) a previous \code{tubeMap} output to add to,
useful if you are plotting lots of different data on the same map and
don't want to have to download the map layer every time.}

\item{...}{Additional arguments. See Notes}
}
\value{
\code{tubeMap} generates static maps using \code{ggplot2}
and \code{OpenStreetMap}. \code{leafletTubeMap} generates interactive
(HTML) maps using \code{leaflet}.
}
\description{
\code{DTEval} currently provides static \code{ggplot2} and
dynamic (html) \code{leaflet} diffusion tube maps.
}
\details{
In addition to \code{data}, the main data source for plots,
\code{ggplotTubeShell} handles the following common plot arguments:
}
\note{
These functions are in development, so may change.

Addition arguments include:
\itemize{
\item \code{point} to add tube locations as points.
\item \code{polygon} to add a polygon to the map.
\item Common plotting arguments, e.g \code{col} for colour, etc. These
typically applied to all plot elements (points, polygons, ect) that
can apply them by can also be specified, e.g. \code{point.col='red'}
to just colour points red.
}
}
