% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.tube.meta.R
\name{test.tube.meta}
\alias{test.tube.meta}
\alias{testTubeMeta}
\alias{tube.meta}
\title{Diffusion Tube Meta Data}
\usage{
testTubeMeta(data, x = NULL, by = NULL, ...)
}
\arguments{
\item{data}{Data source, typically a data.frame or similar, containing
data-series of diffusion tube records.}

\item{x}{(Optional) The name of one or more data-series in \code{data} or
expressions to be evaluated, supplied as character strings. If not supplied,
this defaults to all data-series in \code{code}}

\item{by}{(Optional) The name of one or more data-series or tags to be
used as a case identifier for meta information. If not supplied, this
defaults to the \code{all.data}, \code{sample}, \code{location} and
\code{date} properties of the data set tags. See Details below.}

\item{...}{additional arguments, currently ignored.}
}
\value{
\code{testTubeMeta} generates a plot of meta-like behaviour.
It is a standard \code{ggplot2} output that can be modified using
standard \code{ggplot2} methods.
}
\description{
Testing diffusion tube meta-data. Coded methods are all
currently based on the visualisation of meta-data-like behaviour.
}
\details{
\code{testTubeMeta} generates a plot of meta-like behaviour. It is
currently in-development but is intended to be used in the form:

\code{testTubeMeta(dt.data)}

\code{testTubeMeta} attempts to identify data-series that look like meta
information in the supplied \code{data}. By its nature, meta-data
is expected to unique at a specific level of aggregation. For example,
all samples collected at one site should have the same \code{latitude},
\code{longitude}, sample type, etc regardless of their date of collection.
By default, \code{testTubeMeta} looks for data-series that are meta-like
at four levels of aggregation: \code{all.data} indicating that all records
are identical, \code{sample} indicating that all records for a given
\code{location} and \code{date} combination are identical; \code{location}
indicating that all records for a given \code{location} are identical; and
\code{date} indicating that all records for a given \code{date} are
identical. Any of these that score 100 percentage indicate meta-like
behaviour at the associated level of aggregation. Although, a meta-like does
not prove a data-series is meta-data, less than 100 percent scores for a
data-series that is expected to be meta-data, would indicate some issues
with that data-series.

Although the function
default is to look for known sample identifiers, this grouping term
is set using the addition argument \code{by}. The function then extracts
all data-series with only one unique non-\code{NA} value. By default,
the function test all non-grouping data-series, but similarly
testing/extraction can be limited to specific cases using \code{x}.
Please remeber that this function extracts data-series that at the
\code{by} grouping-level looks like meta information.
}
\examples{
#basic examples
testTubeMeta(dt.brd)

}
\seealso{
\code{\link[=checkTubeMeta]{checkTubeMeta()}} and related functions for working with and
fixing Diffusion Tube meta-data.
}
