% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tube.summaries.R
\name{tube.summaries}
\alias{tube.summaries}
\alias{tubeSummary}
\alias{tubeSummaryLatLon}
\alias{tubeSummarySample}
\title{Miscellaneous Diffusion Tube Code}
\usage{
tubeSummary(data, ...)

tubeSummarySample(data, ...)

tubeSummaryLatLon(data, ...)
}
\arguments{
\item{data}{Data source, typically a data.frame or similar, containing
data-series of diffusion tube records.}

\item{...}{additional arguments, currently ignored.}
}
\value{
All functions return a \code{data.frame} of the requested
\code{data} summary statistics.
}
\description{
Diffusion Tube (DT) data Summaries.
}
\details{
\code{tubeSummary} attempts to generate an overview summary for the
supplied \code{DTEval} diffusion tube data set \code{data}.

\code{tubeSummarySample} attempts to generate a summary of
sample-related counts and check sums for \code{data}.

\code{tubeSummaryLatLon} attempts to calculate the distance between
sample locations (unique lat/lon combinations) and the estimated center
of sampling, and ranks these by distance starting with the most-distance.

They are also intended to be used in the form:

\code{tubeSummary(data)}

And all assumes that \code{data} is data \code{DTEval} will
recognise as Diffusion Tube data, so either previously tagged
tube data or data that is tag-able using a default call of
\code{\link{tagTube}}
}
