% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.tube.R
\name{cluster.tube}
\alias{cluster.tube}
\alias{clusterTubeData}
\title{Cluster Diffusion Tube Data}
\usage{
clusterTubeData(
  data,
  tube = ".value",
  by = "site",
  clusters = 2,
  method = 2,
  ...
)
}
\arguments{
\item{data}{Data source, typically a data.frame or similar, containing
data-series of diffusion tube records.}

\item{tube}{The name of the data-series (in \code{data}) to fit a model
to (see \code{model}), typically the DT NO2 concentrations (in ug/m3).}

\item{by}{The name(s) of hierarchical grouping terms. These are used to
sub-sample the data before clustering data.}

\item{clusters}{The number of clusters to extract, default 2.}

\item{method}{The clustering method to apply: 1 (standard cluster);
2 (clustering of correlations); or 3 (clustering on normalised profiles)}

\item{...}{additional arguments, currently ignored.}
}
\value{
All functions return the supplied \code{data.frame}
with attached predicted time-series components: \code{..season},
\code{..deseason}, etc.
}
\description{
Functions for clustering multiple site
diffusion tube (DT).
}
\details{
\code{clusterTubeData} attempts to cluster data in the supplied time-series.

It assigns data to the requested number of clusters.
}
\note{
\code{clusterTubeData} and related functions assume that \code{data}
is a data set \code{DTEval} will recognise as Diffusion Tube data, so either
previously tagged tube data or data that is tag-able using a
default call of \code{\link{tagTube}}
}
