% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deseason.tube.R
\name{deseason.tube}
\alias{deseason.tube}
\alias{deseasonTubeData}
\title{Deseasonalise Diffusion Tube Data}
\usage{
deseasonTubeData(data, tube = ".value", by = NULL, method = 1, ...)
}
\arguments{
\item{data}{Data source, typically a data.frame or similar, containing
data-series of diffusion tube records.}

\item{tube}{The name of the data-series (in \code{data}) to fit a model
to (see \code{method} and \code{Details}), typically the DT NO2
concentrations (in ug/m3).}

\item{by}{The name(s) of hierarchical grouping terms. These are used to
sub-sample the data when deseasonalising data.}

\item{method}{The deseasonalisation method to apply: 1 (default)
LOESS; 2 local LOESS, like \code{1} but LOESS is applied to
groups of one or more near sites, see also \code{Details}.}

\item{...}{additional arguments, currently passed to \code{loess}.}
}
\value{
All functions return the supplied \code{data.frame}
with attached predicted time-series components: \code{..season},
\code{..deseason}, etc.
}
\description{
Functions for deseasonalising multiple year
diffusion tube (DT) data time-series.
}
\details{
\code{deseasonTubeData} attempts to deseasonalise the supplied time-series.

It builds a LOESS time-series model in the form:

\code{[tube] ~ loess([day-of-year] + [date])}

And, extracts the seasonal component as the response term for the day-of-year.

\code{method 1} applies the LOESS model directly at the requested \code{by} level.

\code{method 2} applies the LOESS model by-location in addition to the requested
\code{by} level. Used directly it is equivalent to:

\code{deseasonTubeData(..., method=1, by=".location")}

... or similar, but can to expanded to build near-location models
using \code{max.distance} and/or \code{max.n} to set the maximum distance
between and/or number of sampling sites in the subset near data used to
build these models.
}
\note{
\code{deseasonTubeData} and related functions assume that \code{data}
is a data set \code{DTEval} will recognise as Diffusion Tube data, so either
previously tagged tube data or data that is tag-able using a
default call of \code{\link{tagTube}}
}
\examples{
# ambient data exhibits seasonality
tubePlot(dt.brd, ".date",".value", plot.type="smooth")
# underlying trend
mod<-deseasonTubeData(dt.brd,".value", by=".location")
tubePlot(mod,".date", "..trend", plot.type="smooth")
}
