% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DTEval-package.R
\name{DTEval-package}
\alias{DTEval-package}
\alias{DTEval}
\title{DTEval: Diffusion Tube Evaluation Code}
\description{
R code used for the pre-processing, analysis and evaluation of diffusion tube (DT)
data typically collected in air quality assessment exercises.

This package contains code developed was part of on-going with
Defra, UK Local Authorities, Ipsos MORI and University of Leeds.

"_PACKAGE"
}
