% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.tube.meta.R
\name{misc.tube.meta}
\alias{misc.tube.meta}
\alias{addTubeMeta}
\alias{checkTubeMeta}
\alias{extractTubeMeta}
\alias{extractAndAddTubeMeta}
\alias{padTubeMeta}
\alias{repairTubeMeta}
\title{Working with Diffusion Tube Meta Data}
\usage{
addTubeMeta(data, ref = NULL, by = NULL, ...)

checkTubeMeta(data, x = NULL, by = NULL, ...)

extractTubeMeta(data, x = NULL, by = NULL, ...)

extractAndAddTubeMeta(data, x = NULL, by = NULL, ref = NULL, ...)

padTubeMeta(data, x = NULL, by = NULL, ...)

repairTubeMeta(data, x = NULL, by = NULL, options = NULL, ...)
}
\arguments{
\item{data}{Data source, typically a data.frame or similar, containing
data-series of diffusion tube records.}

\item{ref}{(For \code{addTubeMeta} and \code{extractAndAddTubeMeta} only)
The meta data source, typically a data.frame or similar containing
meta-information associated with the diffusion tube records in \code{data}.
See Details below.}

\item{by}{The name of a data-series that can be used as a
case identifier for meta information. For most diffusion tube meta data,
this is often a site identifier, e.g. a site code or name. If \code{by}
is not supplied, the function will check for known \code{DTEval}
\code{tags} and data set details, but may still require user input. See
Details below.}

\item{...}{additional arguments, currently ignored.}

\item{x}{(For most \code{...TubeMeta} functions) The name of a data-series in
\code{data} or expression to be evaluated, supplied as a character string.
See Details below.}

\item{options}{(for \code{repairTubeMeta} only) a vector of valid options
for \code{x}. These are used to a guide when attempting to repair
\code{x}/\code{by} combinations with multiple values associated. See Details
below}
}
\value{
These functions are generally intended to be used in the form:

\code{updated.data <- addTubeMeta(dt.data, ref, "[by.name]")}

\code{requested.data <- checkTubeMeta(dt.data, "[meta.name]", "[by.name]")}

... etc

\code{addTubeMeta} returns the supplied \code{data} with \code{ref}
merged at the requested level.

\code{checkTubeMeta} returns a \code{data.frame} of values associated with
found when grouping \code{x} data at the requested level. More than one
value and/or unexpected values indicate that \code{x} may not be meta-data
or, if it is, that the meta-data has been corrupted.

\code{extractTubeMeta} returns a \code{data.frame} of data that looks like
meta-data when grouped at the requested level. (NB: data-series have to be
unique at the grouping level requested to generate an output with
\code{extractTubeMeta}. So, any \code{x}/\code{by} combinations that
\code{checkTubeMeta} reports as having multiple outputs will be ignored by
this function.)

\code{padTubeMeta} returns \code{data} with the requested fix,
if it can be applied.

\code{repairTubeMeta} like \code{padTubeMeta}, returns the fixed \code{data},
if the repair can be made.
}
\description{
Miscellaneous code used to work with
diffusion tube (DT) meta data.
}
\details{
\code{addTubeMeta} attempts to merge \code{data} and \code{ref} using
\code{by} as the common merging-term. It is intended for use with reference
meta data or data extracted with from a reliable source using
\code{extractTubeMeta}.

\code{checkTubeMeta} attempts to identified all unique values for an
\code{x}/\code{by} combination. Meta-data is expected to unique at a
specific level of aggregation, e.g. all \code{latitude} records should
be identical for a given sample site. So, finding multiple values
and/or unexpected values as options for \code{x} suggests the data-series
is not meta-data at that level or it has been corrupted. (See also
\code{repairTubeMeta} below.)

\code{extractTubeMeta} attempts to extract data that looks like meta
information from the supplied \code{data}.
Although the function default is to look for known sample identifiers,
this grouping term can be set using the addition argument \code{by}.
The function then extracts all data-series with only one unique
non-\code{NA} value for case of \code{by}. By default, the function
tests all non-grouping data-series, but similarly testing/extraction
can be limited to specific data-series using \code{x}. Please remember that
this function extracts data-series that looks like meta information at
the \code{by} grouping-level. There is no unambiguous test to identify a
data-series as meta-data. So, this function needs to be handled with care,
especially if you have any concerns about the quality of your data sets.

\code{extractAndAddTubeMeta} is a wrapper for \code{extractTubeMeta} and
\code{addTubeMeta} to extract valid meta-information from a source,
\code{ref}, and add that meta-information to \code{data}. similarly, care
should be taken using this function.

\code{padTubeMeta} attempts to pad the \code{x} data-series by replacing
any \code{NA}s with the first non-\code{NA} entry from the same data-series.
It is intended for use with meta-data data-series, e.g. a \code{latitude}
(or \code{longitude column}) where value was only entered once. Again, care
should be taken using this function.

\code{repairTubeMeta} attempts to repair what looks like bad
meta-information. It uses an extra argument \code{options}, a list of
valid options for \code{x}, as a reference to rationalise multiple value
\code{by} subsets.
}
