% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.dt.fixes.R
\name{misc.dt.fixes}
\alias{misc.dt.fixes}
\alias{padTubeMeta}
\title{Miscellaneous Diffusion Tube Data Fixes}
\usage{
padTubeMeta(data, x = NULL, site.id = NULL, ...)
}
\arguments{
\item{data}{Data source, typically a data.frame or similar, containing
data-series of diffusion tube records.}

\item{x}{The name of a data-series in \code{data} or expression to
be evaluated, supplied as a character string. See Details below.}

\item{site.id}{The name of a data-series that can be used as a
sampling site identifier. If \code{site.id} is not supplied, the
function will check for known \code{DTEval} \code{tags}.
See Details below.}

\item{...}{additional arguments, currently ignored.}
}
\value{
\code{padTubeMeta} returns \code{data} with the requested fix,
if it can be applied.
}
\description{
Miscellaneous code for fixing commonly reported issues with
diffusion tube (DT) data.
}
\details{
\code{padTubeMeta} attempts to pad the \code{x} data-series by replacing
any \code{NA}s with the non-\code{NA} entries from the same data-series.
It is intended for use with meta-data data-series, e.g. a \code{latitude}
(or \code{longitude column}) where value was only entered once. Meta-data
is only be expected to unique at a specific level of aggregation, e.g.
\code{latitude} will be unique for a sample site. So, an additional
identifier is required to group the data, \code{site.id}.

The function returns the supplied data with the attempted fix applied,
and is generally used in the form:

\code{dt.data.new <- padTubeMeta(dt.data, "[meta.name]", "[id.name]")}
}
