% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tag.tube.data.R
\name{tag.tube.data}
\alias{tag.tube.data}
\alias{tagTube}
\alias{tagTubeDates}
\alias{tagTubeLatLon}
\alias{tagTubeSampleID}
\title{Diffusion Tube Pro-processing}
\usage{
tagTube(data, ...)

tagTubeDates(data, method = -1, force = FALSE, ...)

tagTubeLatLon(data, method = -1, force = FALSE, ...)

tagTubeSampleID(data, method = -1, force = FALSE, ...)
}
\arguments{
\item{data}{Data source, typically a
data.frame or similar, containing
data-series of diffusion tube records.}

\item{...}{Additional arguments can be used to assign data sources if
these are not automatically identified by the \code{tag...} functions.
See below for further details.}

\item{method}{The method to apply when testing or modifying \code{data}.
The default (\code{method=-1}) means try all methods in order until one
works. See below for further details on different function methods.}

\item{force}{(logical) By default, these functions do not overwrite
existing tags if they are already present. The option \code{force=TRUE}
forces them to rebuild the requested tag using current setting.}
}
\value{
These functions typically return the supplied \code{data}
modified as requested.
}
\description{
Pre-processing diffusion tube (DT) data for use with
\code{DTEval}. Coded methods to standardise DT data collected using
different DT sampling procedures or different logging conventions.
Tagging makes a version of your tube information in a format
\code{DTEval} can more easily work with, while retaining your original
data in the format you are familiar with.

All \code{tag...} functions are intended to be run in the form:

\code{tagged.data <- tagTube(data, ...)}

(where ... refers to any additional tagging arguments)
}
\details{
All \code{tag...} functions allow you to change information sources when
adding tags. \code{DTEval} has defaults it is expects, but your files
might be different. So, you can identify the specific \code{data} columns
you would like used using additional arguments, e.g.:
\itemize{
\item \code{year} to assign the calendar year
\item \code{month} to assign the calendar month
\item \code{start} and \code{end} to assign the individual DT
sampling period directly (\code{start} deployment and \code{end}
collection, respectively). \code{start} and \code{end} are assumed
to be provided in conventional 'YYYY-MM-DD' format, but this can also
be modified using \code{data.format} and \code{\link{strptime}}
formatting instructions.
\item \code{lat} and \code{lon}  to assign latitude and longitude, respectively.
}

So, for example:

\code{tagTubeLatLon(data, lat="lat", lon="long")}

tells \code{tagTubeLatLon} to use the \code{data} columns named 'lat' and
'long' as the sources when tagging latitude and longitude, respectively.

Function Methods:

By default, these methods try all available methods in order, until they
find a method that works and they run out methods. If you want to use a
specific method, or method order, just specify them, e.g. to tag tube
dates first trying \code{method} 3 then 2:

\code{tagTubeDates(data, method=3:2)}

\code{tagTubeDates} attempts to build \code{Date} class time-stamps for
sampling dates. It currently has three methods: \code{method=1} use month
and year records to build a nominal calendar monthly sampling record,
i.e., first to last day of each month; \code{method=2} use month and year
records to build a sampling calendar based on the DEFRA diffusion tube
sampling calendar (see \code{\link{dt.calendar}}); and, \code{method=3}
to build directly using user-assigned sampling \code{start} and
\code{end} dates. If sampling dates are built, these are added to
\code{data} as columns \code{.start_date} and \code{.end_date}.

\code{tagTubeLatLon} identifies sample location. It currently uses
one method: \code{method=1} assigning all locations based on identified
latitude and longitude. These are then tagged as \code{.latitude} and
\code{.longitude}, respectively.

\code{tagTubeSampleID} identifies samples. It currently uses
one method: \code{method=1} assigning all same-location-and-time DTs as
replicates, and is added to \code{data} as \code{.sample_id}.

\code{tagTube} is a wrapper that runs \code{tagTubeDates},
\code{tagTubeLatLon} and \code{tagTubeiteID}. Other \code{DTEval}
functions typically pass your \code{data} to \code{tagTube} to check for
tags and add them if missing. If you are using \code{tagTube} or one of
these on untagged data and need to specific individual methods for each
\code{tagTube...} function, prefix the method request, e.g. to use method
2 for \code{tabTubeDates} but leave the other functions using default
setting when applying \code{tagTube}, use:

\code{tagTube(data, dates.method=2)}
}
