% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tube.maps.R
\name{tube.maps}
\alias{tube.maps}
\alias{tubeMap}
\alias{leafletTubeMap}
\title{Diffusion Tube maps}
\usage{
tubeMap(data, x = NULL, y = NULL, ...)

leafletTubeMap(data, x = NULL, y = NULL, plot.type = NULL, ...)
}
\arguments{
\item{data}{Data source to be used to build a plot. This is typically a
data.frame or similar, but can also be a previous plot generated by the
same function.}

\item{x, y}{The names of the data-series to plot on the
X and Y axes, respectively, and assumed to be sampling point longitude and
latitude.}

\item{...}{Additional arguments. \code{tubeMap} uses \code{data}, \code{x},
\code{y} and optional argument \code{facet} to build map layers then passes
arguments on to \code{tubePlot} to generate plot layers.}

\item{plot.type}{The type of plot(s) to produce, e.g.:
\code{'point'} (default) for a scatter plot, \code{'surface'} to
fit a surface, etc...}
}
\value{
\code{tubeMap} generates static maps using \code{ggplot2}
and \code{OpenStreetMap}. \code{leafletTubeMap} generates interactive
(HTML) maps using \code{leaflet}.
}
\description{
\code{DTEval} currently provides static \code{ggplot2} and
dynamic (html) \code{leaflet} diffusion tube maps.
}
\note{
These functions are in development, so may change.
}
\examples{
\dontrun{
# map layers require internet access

#basic examples
tubeMap(dt.brd)
leafletTubeMap(dt.brd)
}

}
