% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tube.maps.R
\name{tube.maps}
\alias{tube.maps}
\alias{tubeMap}
\alias{leafletTubeMap}
\title{Diffusion Tube maps}
\usage{
tubeMap(data, x = NULL, y = NULL, ...)

leafletTubeMap(data, x = NULL, y = NULL, ...)
}
\arguments{
\item{data}{Data source to be used to build a plot. This is typically a
data.frame or similar, but can also be a previous plot generated by the
same function.}

\item{x, y}{The names of the data-series to plot on the
X and Y axes, respectively, and assumed to be sampling point longitude and
latitude.}

\item{...}{Additional arguments. See \code{Details}.}
}
\value{
\code{tubeMap} generates static maps using \code{ggplot2}
and \code{OpenStreetMap}. \code{leafletTubeMap} generates interactive
(HTML) maps using \code{leaflet}.
}
\description{
\code{DTEval} currently provides static \code{ggplot2} and
dynamic (html) \code{leaflet} diffusion tube maps.
}
\details{
In addition to \code{data}, \code{x} and \code{y}, \code{tubeMap}
handles the following arguments:
\itemize{
\item \code{point} to add tube locations as points.
\item \code{polygon} to add a polygon to the map.
\item \code{surface} to fit/add a surface to the the map.
\item Common plotting arguments, e.g \code{col} for colour, etc. These are
typically applied to all plot elements (points, polygons, ect) that
can apply them but plot features can also be specified, e.g. \code{point.col='red'}
to just colour points red.
}
}
\note{
These functions are in development, so may change.
}
