% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tube.maps.R
\name{tube.maps}
\alias{tube.maps}
\alias{tubeMap}
\alias{tubeMap_leaflet}
\alias{tubeMap_ggplot2}
\title{Diffusion Tube maps}
\usage{
tubeMap(data, x = NULL, y = NULL, polygon = NULL, plot.type = "leaflet", ...)

tubeMap_leaflet(
  data,
  x = NULL,
  y = NULL,
  polygon = NULL,
  plot.type = "point",
  ...
)

tubeMap_ggplot2(
  data,
  x = NULL,
  y = NULL,
  polygon = NULL,
  plot.type = "point",
  ...
)
}
\arguments{
\item{data}{Data source, typically a data.frame or similar, to be used to
build a plot using ggplot2.}

\item{x, y}{The names of the data-series to plot on the
X and Y axes, respectively.}

\item{polygon}{(Optional) Polygon of boundary records, e.g.
CAZ boundaries.}

\item{plot.type}{Type of plot generated, current options are 'leaflet'
(default) and 'ggplot2'.}

\item{...}{Additional arguments. See Notes}
}
\value{
\code{tubeMap_leaflet} and \code{tubeMap_ggplot2} return
\code{leaftlet} and \code{ggplot2} maps, respectively. \code{tubeMap}
is wrapper for both.
}
\description{
\code{DTEval} currently provides static \code{ggplot2} and
dynamic (html) \code{leaflet} diffusion tube maps.
}
\details{
In addition to \code{data}, the main data source for plots,
\code{ggplotTubeShell} handles the following common plot arguments:
}
\note{
These functions are in development, so may change.
}
