% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tube.plots.R
\name{tube.plots}
\alias{tube.plots}
\alias{tubePlot}
\alias{tubeTimePlot}
\alias{ggplotTubeShell}
\title{Diffusion Tube Data Plots}
\usage{
tubePlot(data, x = NULL, y = NULL, plot.type = "point", ...)

tubeTimePlot(data, x = NULL, y = NULL, plot.type = "point", ...)

ggplotTubeShell(data, x = NULL, y = NULL, ...)
}
\arguments{
\item{data}{Data source, typically a data.frame or similar, to be used to
build a plot using ggplot2.}

\item{x, y}{The names of the data-series to plot on the
X and Y axes, respectively, and assumed to be elements of \code{data}.}

\item{plot.type}{The type of plot required.}

\item{...}{Additional arguments. See details below.}
}
\value{
\code{tubePlot} is a general purpose plotting functions for
diffusion tube data.

\code{tubeTimePlot} is a \code{tubePlot} wrapper that adds the tagged
sampling date as a second axes if one of \code{x} or \code{y} is
supplied.

\code{ggplotTubeShell} returns a ggplot object the user then
needs to add \code{geom}s to, to build a standard plot.
}
\description{
\code{DTEval} uses \code{ggplot2} to generate most plots.

Obviously, you are welcome to use any graphics package you know to generate
your own plot. In fact, if you want really fine control of plot outputs, we
recommend taking the time to learn a graphics package like
\code{ggplot}, \code{lattice}, etc. But \code{DTEval} also includes
several quick-use wrappers for some commonly used plot types and useful
options like grouping and faceting, which can really help with data
visualisation. It also includes \code{ggplotTubeShell}, which builds
common plot shells for many of these, and handles some of of
the generic plot control for functions like \code{\link{testTubePrecision}}
and \code{\link{testTubePrecision}}.
}
\details{
In addition to \code{data}, the main data source for plots,
\code{ggplotTubeShell} handles the following common plot arguments:
\itemize{
\item \code{xlab}, \code{ylab} The X and Y axes labels to use if
different from plot defaults.
\item \code{group} The name of the data-series to use to group
plotted data with. By default, \code{ggplotTubeShell} assumes
supplied groups are factors if less than 20 unique cases.
\item \code{facet} The name of the data-series to use to cut
the data by when generating multiple plot panels. Also provides
an addition \code{facet.type} shortcut to different \code{ggplot2}
facet options: \code{'wrap'} for \code{facet_wrap} (default),
\code{'grid'} for \code{facet_grid} (or 'grid.row' or grid.col' to
specific by row-then-column or column-then-row grid handling,
respectively).
}

Functions like \code{tubePlot} add layers to this framework to
generate common plots.
}
\note{
Maybe I'm already missing \code{qplot}...
}
